Revamp chat room page layout
