--- pynicotine/gtkgui/ui/roomlist.ui
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<interface>
-  <requires lib="gtk+" version="3.18"/>
-  <object class="GtkBox" id="vbox2">
-    <property name="visible">1</property>
-    <property name="orientation">vertical</property>
-    <child>
-      <object class="GtkBox">
-        <property name="visible">1</property>
-        <property name="margin-start">6</property>
-        <property name="margin-end">6</property>
-        <property name="margin-top">6</property>
-        <property name="margin-bottom">6</property>
-        <property name="spacing">6</property>
-        <child>
-          <object class="GtkLabel" id="ChatroomsLabel">
-            <property name="visible">1</property>
-            <property name="halign">start</property>
-            <property name="label" translatable="yes">&lt;b&gt;Chat Rooms&lt;/b&gt;</property>
-            <property name="use-markup">1</property>
-            <property name="margin-start">5</property>
-            <property name="margin-end">10</property>
-          </object>
-        </child>
-        <child>
-          <object class="GtkSearchEntry" id="SearchRooms">
-            <property name="visible">1</property>
-            <property name="can-focus">1</property>
-            <property name="hexpand">1</property>
-            <property name="width-chars">15</property>
-            <signal name="activate" handler="on_search_room" swapped="no"/>
-          </object>
-        </child>
-        <child>
-          <object class="GtkButton" id="RefreshButton">
-            <property name="visible">1</property>
-            <property name="can-focus">1</property>
-            <property name="tooltip-text" translatable="yes">Refresh room list</property>
-            <signal name="clicked" handler="on_refresh" swapped="no"/>
-            <child>
-              <object class="GtkImage">
-                <property name="visible">1</property>
-                <property name="icon-name">view-refresh-symbolic</property>
-              </object>
-            </child>
-            <style>
-              <class name="circular"/>
-              <class name="image-button"/>
-            </style>
-          </object>
-        </child>
-      </object>
-    </child>
-    <child>
-      <object class="GtkScrolledWindow">
-        <property name="visible">1</property>
-        <property name="can-focus">1</property>
-        <property name="vexpand">1</property>
-        <child>
-          <object class="GtkTreeView" id="RoomsList">
-            <property name="visible">1</property>
-            <property name="can-focus">1</property>
-          </object>
-        </child>
-      </object>
-    </child>
-    <child>
-      <object class="GtkBox">
-        <property name="visible">1</property>
-        <property name="margin-start">8</property>
-        <property name="margin-end">8</property>
-        <property name="margin-top">8</property>
-        <property name="margin-bottom">8</property>
-        <property name="orientation">vertical</property>
-        <child>
-          <object class="GtkCheckButton" id="AcceptPrivateRoom">
-            <property name="label" translatable="yes">Accept private room invitations</property>
-            <property name="visible">1</property>
-            <property name="can-focus">1</property>
-            <property name="use-underline">1</property>
-          </object>
-        </child>
-      </object>
-    </child>
-  </object>
-</interface>
